+++ src/main/java/org/example/domain/Product.java
@@ -4,16 +4,16 @@ public abstract class Product {
 
     private int id;
     private String name;
-    private int price;
+    private double price;
     private int quantity;
 
-    public Product(String name, int price, int quantity) {
+    public Product(String name, double price, int quantity) {
         this.price = price;
         this.name = name;
         this.quantity = quantity;
     }
 
-    public Product(int id, String name, int price, int quantity) {
+    public Product(int id, String name, double price, int quantity) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -32,7 +32,7 @@ public abstract class Product {
         return quantity;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 
@@ -44,7 +44,7 @@ public abstract class Product {
         this.quantity = quantity;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         this.price = price;
     }
